add betterbib/sync.py
