+++ Makefile
@@ -23,6 +23,9 @@ clean:
 	@find . | grep -E "(__pycache__|\.pyc|\.pyo$\)" | xargs rm -rf
 	@rm -rf *.egg-info/ build/ dist/ betterbib_cache.sqlite/
 
+black:
+	black setup.py betterbib/ test/
+
 lint:
 	black --check setup.py betterbib/ test/
 	flake8 setup.py betterbib/ test/
